refactor: split oversized backtesting test modules
- simulate.rs, walk_forward.rs, and the portfolio tests grew past
  the 800-line file cap on this branch; test modules move to sibling
  files with identical test paths
